client/archival-material-client/src/hooks/useOpenSeaDragonFilters/useOpenSeaDragonFilters.test.ts
export {}
/*import { renderHook, act } from '@testing-library/react-hooks';
import { useOpenSeadragonContext } from '../../context/OpenSeadragonContext';
import { useOpenSeadragonFilters } from './useOpenSeaDragonFilters';

jest.mock('../../context/OpenSeadragonContext', () => ({
  useOpenSeadragonContext: jest.fn(),
}));

describe('useOpenSeadragonFilters', () => {
  beforeEach(() => {
    (useOpenSeadragonContext as jest.Mock).mockReturnValue({
        viewer: {
          canvas: {
            style: {},
          },
        },
      });
  });

  it('should update brightness', () => {
    const { result } = renderHook(() => useOpenSeadragonFilters());
    act(() => {
      result.current.setBrightness(50);
    });
    expect(result.current.brightness).toBe(50);
  });

  it('should update contrast', () => {
    const { result } = renderHook(() => useOpenSeadragonFilters());
    act(() => {
      result.current.setContrast(75);
    });
    expect(result.current.contrast).toBe(75);
  });

  // Add similar tests for other filter functions

  it('should reset filters', () => {
    const { result } = renderHook(() => useOpenSeadragonFilters());
    act(() => {
      result.current.setBrightness(50);
      result.current.setContrast(75);
      result.current.resetFilters();
    });
    expect(result.current.brightness).toBe(100);
    expect(result.current.contrast).toBe(100);
    // Add assertions for other filters being reset to their default values
  });

  // Add more test cases as needed for other functionality
});
*/